# src/components/Swagger/api-definition.yaml
openapi: 3.0.0
info:
  title: Calcetto API
  description: <strong><i>API definitions for Calcetto API server</i></strong>
  version: 1.0.0-oas3
# Server
servers:
  - url: "http://localhost:5001/api/v2/"
    description: Development server.

# Tags
tags:
  - name: Authentication
    description: Manage user, Login, Logout or Check authentication status
  - name: Tournament
    description: Manage tournament
  - name: Players
    description: Manage players

# Path
paths:
  #-----------------------------------------------------------------------------
  # Authentication
  #
  "/auth/check":
    get:
      tags:
        - Authentication
      summary: Check if user is already logged.
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationReponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"

  "/auth/logout":
    get:
      summary: Logout.
      tags:
        - Authentication
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationReponse"

  "/auth/update":
    put:
      summary: |
        Update existing user and related player if exists.
      tags:
        - Authentication
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user:
                  $ref: "#/components/schemas/UserDTO"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationReponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

  "/auth/login":
    post:
      summary: Authenticate user
      tags:
        - Authentication
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:
                username: User
                password: UserPasswordInPlainText
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationReponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

  "/auth/register":
    post:
      summary: Register new user
      tags:
        - Authentication
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                name:
                  type: string
                surname:
                  type: string
                email:
                  type: string
                cEmail:
                  type: string
                password:
                  type: string
                cPassword:
                  type: string
                phone:
                  type: string
                birthday:
                  type: string
                playerRole:
                  $ref: "#/components/schemas/PlayerRole"
              example:
                username: "username"
                name: "name"
                surname: "surname"
                email: "test@email.com"
                cEmail: "test@email.com"
                password: "verylongpassword"
                cPassword: "verylongpassword"
                phone: "3331234598"
                birthday: ""
                playerRole: "GoalKeeper"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationReponse"
        "400":
          description: Found errors
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BasicResponse"
                  - type: object
                    properties:
                      errors:
                        type: array
                        items:
                          type: string
        "500":
          $ref: "#/components/responses/ServerError"

  "/auth/delete":
    delete:
      summary: Delete user
      tags:
        - Authentication
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationReponse"
        "400":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BasicResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

  #-----------------------------------------------------------------------------
  # Tournament
  #
  "/tournament/list":
    get:
      tags:
        - Tournament
      summary: List all tournament
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BasicResponse"
                  - type: object
                    properties:
                      tournamentsList:
                        type: array
                        items:
                          $ref: "#/components/schemas/TournamentDTO"
        "500":
          $ref: "#/components/responses/ServerError"

  "/tournament/{tId}":
    get:
      tags:
        - Tournament
      summary: List all tournament
      parameters:
        - in: path
          name: tId
          schema:
            type: integer
          required: true
          description: Numeric ID of the tournament to get
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BasicResponse"
                  - type: object
                    properties:
                      tournamentsList:
                        type: array
                        items:
                          $ref: "#/components/schemas/TournamentDTO"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

  "/tournament/update":
    put:
      summary: |
        Update existing tournament.
      tags:
        - Tournament
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                tournament:
                  $ref: "#/components/schemas/TournamentDTO"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BasicResponse"
                  - type: object
                    properties:
                      tournament:
                        $ref: "#/components/schemas/TournamentDTO"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

  "/tournament/new":
    post:
      summary: |
        Create new tournament.
      tags:
        - Tournament
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                model:
                  $ref: "#/components/schemas/TournamentDTO"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BasicResponse"
                  - type: object
                    properties:
                      tournament:
                        $ref: "#/components/schemas/TournamentDTO"
        "500":
          $ref: "#/components/responses/ServerError"

  #-----------------------------------------------------------------------------
  # Player
  #
  "/player/list":
    get:
      tags:
        - Players
      summary: List all players
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BasicResponse"
                  - type: object
                    properties:
                      playersList:
                        type: array
                        items:
                          $ref: "#/components/schemas/PlayerDTO"
        "500":
          $ref: "#/components/responses/ServerError"

  "/player/list/{tId}":
    get:
      tags:
        - Players
      summary: List all players in a tournament
      parameters:
        - in: path
          name: tId
          schema:
            type: integer
          required: true
          description: Numeric ID of the tournament
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BasicResponse"
                  - type: object
                    properties:
                      playersList:
                        type: array
                        items:
                          $ref: "#/components/schemas/PlayerDTO"
        "500":
          $ref: "#/components/responses/ServerError"

  "/players/update":
    put:
      summary: |
        Update an existing player.
      tags:
        - Players
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                player:
                  $ref: "#/components/schemas/PlayerDTO"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BasicResponse"
                  - type: object
                    properties:
                      player:
                        $ref: "#/components/schemas/PlayerDTO"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

  "/players/new":
    post:
      summary: |
        Create new tournament.
      tags:
        - Players
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                model:
                  $ref: "#/components/schemas/PlayerDTO"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BasicResponse"
                  - type: object
                    properties:
                      player:
                        $ref: "#/components/schemas/PlayerDTO"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"

  "/players/delete":
    delete:
      summary: Delete players
      tags:
        - Players
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BasicResponse"
        "500":
          $ref: "#/components/responses/ServerError"
#-----------------------------------------------------------------------------
# Components
#
components:
  #---------------------------------------------------------------------------
  # Schema : $ref: '#/components/schemas/{name}'
  #
  schemas:
    BasicResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: HTTPStatusCode
        message:
          type: string
        userMessage:
          $ref: "#/components/schemas/UserMessage"
    UserMessage:
      type: object
      required:
        - type
        - message
      properties:
        type:
          type: string
          enum:
            - danger
            - warning
            - success
        message:
          type: string
    AuthenticationReponse:
      allOf:
        - $ref: "#/components/schemas/BasicResponse"
        - type: object
          properties:
            user:
              $ref: "#/components/schemas/UserDTO"
    UserDTO:
      type: object
      additionalProperties: false
      required:
        - id
        - username
        - name
        - surname
        - email
        - phone
        - role
        - birthday
        - label
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        surname:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        birthday:
          type: string
          description: It's a javascript Date object
        label:
          type: string
        player:
          $ref: "#/components/schemas/PlayerDTO"
    UserRole:
      type: string
      enum:
        - Admin
        - User
    PlayerDTO:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - surname
        - alias
        - label
        - role
        - email
        - phone
        - editable
        - rowNumber
      properties:
        id:
          type: integer
        name:
          type: string
        surname:
          type: string
        alias:
          type: string
        label:
          type: string
        role:
          $ref: "#/components/schemas/PlayerRole"
        email:
          type: string
        phone:
          type: string
        userId:
          type: string
          description: |
            Reference to UserDTO
        match_played:
          type: integer
        match_won:
          type: integer
        total_score:
          type: integer
        editable:
          type: boolean
          description: |
            Indicates if player can be deleted.
            A player can be deleted if is not part of any pairs. ( see PairDTO )
        rowNumber:
          type: integer
          description: Used by FE to show row number on table
    PlayerRole:
      type: string
      enum:
        - Non sono un giocatore
        - Portiere
        - Master
        - Attaccante
    TournamentDTO:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - date
        - progress
        - public
        - autoOrder
        - label
        - ownerId
      properties:
        id:
          type: integer
        name:
          type: string
        date:
          type: string
        progress:
          $ref: "#/components/schemas/TournamentProgress"
        public:
          type: boolean
        autoOrder:
          type: boolean
        label:
          type: string
        ownerId:
          type: integer

    TournamentProgress:
      type: string
      enum:
        - new
        - pairsSelection
        - stage1
        - stage2

  #---------------------------------------------------------------------------
  # Responses : $ref: '#/components/responses/{name}'
  #
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BasicResponse"
    NotFound:
      description: Entity not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BasicResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BasicResponse"
    ServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BasicResponse"
  #---------------------------------------------------------------------------
  # Parameters : $ref: '#/components/parameters/{name}'
  #
  # parameters:
  #  offsetParam:
  #    name: offset
  #    in: query
  #    description: Number of items to skip before returning the results.
  #    required: false
  #    schema:
  #      type: integer
  #      format: int32
  #      minimum: 0
  #      default: 0
  #